handle trying to move a filestore that doesn't actually exist on the filesystem yet.

+++ subprojects/core/src/main/groovy/org/gradle/api/internal/filestore/PathKeyFileStore.java
@@ -76,10 +76,12 @@ public class PathKeyFileStore implements FileStore<String>, FileStoreSearcher<St
     }
 
     public void moveFilestore(File destination) {
-        try {
-            FileUtils.moveDirectory(baseDir, destination);
-        } catch (IOException e) {
-            throw new UncheckedException(e);
+        if (baseDir.exists()) {
+            try {
+                FileUtils.moveDirectory(baseDir, destination);
+            } catch (IOException e) {
+                throw new UncheckedException(e);
+            }
         }
 
         baseDir = destination;
